+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Unreleased
+- Defines the internal `_events` property as [non-enumerable][for-in].
+
+[for-in]: http://www.ecma-international.org/ecma-262/5.1/#sec-12.6.4
+
 ## 0.1.338 (Jan 5, 2014)
 - Allows using event names that exist on `Object.prototype`.